findall: DeedInfo -> BidInfo, index my sender+seal.
Strictly speaking, there's no need to retrieve the deed address
(to be indexed by) and the deed's info, since all that's needed
to cancel a bid is the original bidder's address and the bid.

That's what the registrar indexes bids by anyway.

Not trying to retrieve the actual deed saves us a web3 call, which
would quite likely become a bottleneck if checked for _every_
bid.

After all, most bids _will_ be revealed, so it's only useful to
check the actual bid right before issuing the cancel tx. And that's
not the job of this script.