get-stats: specify historic block to look up deed size.
Requires a full sync from the .eth Registrar launch and onwards. :/

Looking up deed size from event-creating tx is not fool-proof. The
tx might, for example, be to a proxy contracts that bids on several
names, "all or nothing", or something similar.

The bottom line is, the only reliable way to get the Deed size of
a specific Deed is to literally _look up its historic balance_.
Everything else has too many edge cases to handle.